Fix various compiler warnings (bug #322)
git-svn-id: https://origsvn.digium.com/svn/asterisk/trunk@1570 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ apps/app_queue.c
@@ -1384,7 +1384,7 @@ static int queues_show(int fd, int argc, char **argv)
 			ast_cli(fd, "   Callers: \n");
 			for (qe = q->head; qe; qe = qe->next) 
 				ast_cli(fd, "      %d. %s (wait: %ld:%2.2ld)\n", pos++, qe->chan->name,
-								(now - qe->start) / 60, (now - qe->start) % 60);
+								(long)(now - qe->start) / 60, (long)(now - qe->start) % 60);
 		} else
 			ast_cli(fd, "   No Callers\n");
 		ast_cli(fd, "\n");
